fix: 背景 git 操作失敗或逾時不再讓 auto-refresh 永久卡死 (#83)
- pending_refresh 從單向閂鎖改成一次性 token：背景 fetch/checkout 失敗時
  只送 NotifyError、不送 AutoRefresh，舊版因此會把 file watcher 永久卡死；
  新版讀了就清，不需要任何呼叫端負責清除
- spawn_git_task 改用有 timeout 的 run_with_timeout（fetch 60s／checkout
  1800s），逾時不再無限期掛住背景 thread
- 抽出 src/process.rs（原本在 github.rs），供 github.rs 與 app.rs 共用
  run_with_timeout
- 背景 git 指令加 GIT_TERMINAL_PROMPT=0／GIT_SSH_COMMAND=ssh -o
  BatchMode=yes／-c gc.auto=0，避免在 raw mode 終端機上卡住等憑證輸入、
  背景意外觸發 gc
