+++ src/event.rs
@@ -371,7 +371,11 @@ pub struct EventController {
     rx: Receiver,
     stop: Arc<AtomicBool>,
     handle: Arc<Mutex<Option<thread::JoinHandle<()>>>>,
-    pending_refresh: Option<Arc<AtomicBool>>,
+    /// 「已有 refresh 在路上」的一次性 token——`mark_pending_refresh` 設它、
+    /// `start_git_watcher` 的背景 thread 用 `swap(false, ...)` 消費，見該處
+    /// 註解。無條件建好（不是 `Option`）：沒有 watcher 時這個 flag 只是沒人
+    /// 讀，不是需要特判的錯誤狀態。
+    pending_refresh: Arc<AtomicBool>,
     term_signal: Arc<AtomicBool>,
     heartbeat: Arc<AtomicU64>,
 }
@@ -419,7 +423,7 @@ impl EventController {
             rx,
             stop: Arc::new(AtomicBool::new(false)),
             handle: Arc::new(Mutex::new(None)),
-            pending_refresh: None,
+            pending_refresh: Arc::new(AtomicBool::new(false)),
             term_signal,
             heartbeat: Arc::new(AtomicU64::new(0)),
         };
@@ -589,31 +593,50 @@ impl EventController {
         self.rx.recv()
     }
 
-    pub fn start_git_watcher(&mut self, repo_root: &Path) {
-        let flag = start_git_watcher(self.tx.clone(), repo_root);
-        self.pending_refresh = Some(flag);
-    }
-
-    pub fn clear_pending_refresh(&self) {
-        if let Some(ref flag) = self.pending_refresh {
-            flag.store(false, Ordering::Release);
-        }
+    pub fn start_git_watcher(&self, repo_root: &Path) {
+        start_git_watcher(self.tx.clone(), self.pending_refresh.clone(), repo_root);
     }
 
     /// 標記「已有 refresh 在路上」，讓 watcher 短期內偵測到的後續 fs 事件
     /// 被 debounce 吃掉，避免主動 refresh 後 watcher 重複觸發 slow-path。
+    ///
+    /// 一次性 token：watcher 端消費掉就清掉（見 `claim_send_slot`），不需要任何
+    /// 呼叫端負責清除。就算標記後的操作本身失敗（只送 `NotifyError`、不送
+    /// `AutoRefresh`），watcher 也不會因此永久卡住——最壞情況是多吞一次
+    /// 無關的 fs 事件。
     pub fn mark_pending_refresh(&self) {
-        if let Some(ref flag) = self.pending_refresh {
-            flag.store(true, Ordering::Release);
-        }
+        self.pending_refresh.store(true, Ordering::Release);
     }
 }
 
-pub fn start_git_watcher(tx: Sender, repo_root: &Path) -> Arc<AtomicBool> {
-    use notify_debouncer_mini::new_debouncer;
+/// 節流視窗內、或 `pending` token 被設過，都不送 `AutoRefresh`——後者是
+/// 背景 git 操作（`spawn_git_task`）主動觸發的 refresh 順便產生的 fs 事件，
+/// 不必疊加一次。抽出來獨立測：watcher thread 本身只做管線接線，這個判斷
+/// 才是真正需要單元測試覆蓋的邏輯。
+///
+/// 注意這不是 predicate——呼叫一次就會消費 `pending` token、推進
+/// `last_sent`，語意跟著改變，不能被安全地重複呼叫來「先問再做」。
+///
+/// `pending` 用 `swap(false, ...)` 消費，讀了就清掉——不像單向閂鎖那樣需要
+/// 第三方負責清除，沒有人清得掉的話，一次背景操作失敗就會把 watcher 永久
+/// 卡住。
+fn claim_send_slot(
+    pending: &AtomicBool,
+    now: Instant,
+    last_sent: &mut Instant,
+    throttle: Duration,
+) -> bool {
+    if now.duration_since(*last_sent) < throttle {
+        return false;
+    }
+    // 過了節流視窗就重設時鐘；吞掉的這批也算「已送」，讓視窗蓋住背景操作
+    // 觸發的 fs 事件尾巴，不會緊接著又被下一批事件重新判定成「該送」。
+    *last_sent = now;
+    !pending.swap(false, Ordering::AcqRel)
+}
 
-    let pending_refresh = Arc::new(AtomicBool::new(false));
-    let pending = pending_refresh.clone();
+fn start_git_watcher(tx: Sender, pending: Arc<AtomicBool>, repo_root: &Path) {
+    use notify_debouncer_mini::new_debouncer;
 
     let repo_root = repo_root.to_path_buf();
     let git_dir = repo_root
@@ -658,12 +681,8 @@ pub fn start_git_watcher(tx: Sender, repo_root: &Path) -> Arc<AtomicBool> {
                         continue;
                     }
                     let now = Instant::now();
-                    if now.duration_since(last_sent) < throttle {
-                        continue;
-                    }
-                    if !pending.swap(true, Ordering::AcqRel) {
+                    if claim_send_slot(&pending, now, &mut last_sent, throttle) {
                         tx.send(AppEvent::AutoRefresh);
-                        last_sent = now;
                     }
                 }
                 Ok(Err(_)) => {}
@@ -671,8 +690,6 @@ pub fn start_git_watcher(tx: Sender, repo_root: &Path) -> Arc<AtomicBool> {
             }
         }
     });
-
-    pending_refresh
 }
 
 /// 先走快速路徑：在任何 syscall 之前，先對原始 event path 做便宜的字串檢查。
@@ -1086,6 +1103,56 @@ mod tests {
             "心跳週期 {TICK_INTERVAL:?} 對 stall 門檻 {WATCHDOG_STALL_TIMEOUT:?} 來說太長"
         );
         assert!(WATCHDOG_INTERVAL < WATCHDOG_STALL_TIMEOUT);
+    }
+
+    // ── claim_send_slot() ──
+
+    const THROTTLE: Duration = Duration::from_secs(1);
+
+    /// 節流視窗過了、沒有 pending token：正常送出，且更新 `last_sent`。
+    #[test]
+    fn claim_send_slot_sends_when_not_throttled_and_no_pending_token() {
+        let pending = AtomicBool::new(false);
+        let mut last_sent = Instant::now() - THROTTLE * 2;
+        let now = Instant::now();
+
+        assert!(claim_send_slot(&pending, now, &mut last_sent, THROTTLE));
+        assert_eq!(last_sent, now);
+    }
+
+    /// 節流視窗內：不送，且不觸碰 `pending`（沒有消費掉任何人設的 token）。
+    #[test]
+    fn claim_send_slot_swallows_within_throttle_window_without_consuming_token() {
+        let pending = AtomicBool::new(true);
+        let mut last_sent = Instant::now();
+        let now = last_sent + THROTTLE / 2;
+
+        assert!(!claim_send_slot(&pending, now, &mut last_sent, THROTTLE));
+        assert!(
+            pending.load(Ordering::Acquire),
+            "節流視窗內不該消費 token，留給視窗外的下一次判斷"
+        );
+    }
+
+    /// `pending` 是一次性 token：第一次呼叫吞掉（不送）並清成 `false`，且更新
+    /// `last_sent`；緊接著第二次呼叫（視窗外）沒有 token 可吞，正常送出。
+    /// 這是把單向閂鎖換成消費式 token 的核心不變式：token 讀了就清，不需要
+    /// 任何第三方負責清除。
+    #[test]
+    fn claim_send_slot_consumes_pending_token_exactly_once() {
+        let pending = AtomicBool::new(true);
+        let mut last_sent = Instant::now() - THROTTLE * 2;
+        let first = Instant::now();
+
+        assert!(!claim_send_slot(&pending, first, &mut last_sent, THROTTLE));
+        assert!(!pending.load(Ordering::Acquire), "token 應該被消費掉");
+        assert_eq!(last_sent, first, "吞掉的這批也算已送，更新 last_sent");
+
+        let second = first + THROTTLE * 2;
+        assert!(
+            claim_send_slot(&pending, second, &mut last_sent, THROTTLE),
+            "token 已經被上一次呼叫消費掉，這次沒有東西可吞，該正常送出"
+        );
     }
 
     #[test]
